pyproject.toml
[project]
name = "spen-recons"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "pillow>=12.1.0",
    "scikit-image>=0.26.0",
    "scipy>=1.17.0",
    "spenpy",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]

[tool.uv.sources]
spenpy = { git = "https://github.com/txz32102/spenpy.git" }
